\name{engine}
\alias{engine}
\title{Engine Failure Time Data}
\usage{data(engine)}
\description{
  The \code{engine} data frame has 32 rows and 2 columns.
  The first column contains the corrosion level, the second
  column gives the engine failure time.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Time}{A numeric vector of corrosion levels.}
    \item{Corrosion}{A numeric vector of failure times.}
  }
}
\source{
  Unknown.
}
\keyword{datasets}
