\name{fremantle}
\alias{fremantle}
\title{Annual Maximum Sea Levels at Fremantle, Western Australia}
\usage{data(fremantle)}
\description{
  The \code{fremantle} data frame has 86 rows and 3 columns.
  The second column gives 86 annual maximimum sea levels
  recorded at Fremantle, Western Australia, within the period
  1897 to 1989.
  The first column gives the corresponding years.
  The third column gives annual mean values of the Southern
  Oscillation Index (SOI), which is a proxy for meteorological
  volitility.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Year}{A numeric vector of years.}
    \item{SeaLevel}{A numeric vector of annual sea level
      maxima.}
    \item{SOI}{A numeric vector of annual mean values of
      the Southern Oscillation Index.}
  }
}
\source{
  Coles, S. G. (2001)
  \emph{An Introduction to Statistical Modelling of
    Extreme Values.} London: Springer.
}

\keyword{datasets}
