\name{glass}
\alias{glass}
\title{Breaking Strengths of Glass Fibres}
\usage{data(glass)}
\description{
  A numeric vector containing breaking strengths of 63 glass
  fibres of length 1.5 centimetres, recorded under experimental
  conditions.
}
\format{A vector containing 63 observations.}
\source{
  Smith, R. L. and Naylor, J. C. (1987)
  A comparison of maximum likelihood and Bayesian estimators
  for the three-parameter Weibull distribution.
  \emph{Applied Statistics} \bold{36}, 358--396.
}

\references{
  Coles, S. G. (2001)
  \emph{An Introduction to Statistical Modelling of
    Extreme Values.} London: Springer.
}

\keyword{datasets}
