\name{ismev}
\alias{ismev}
\title{ismev -- The Extremes Toolkit: Weather and Climate Applications of Extreme-Value Statistics}
\description{
\pkg{ismev} includes functions to support the computations carried out in
  `An Introduction to Statistical Modeling of Extreme Values' by
  Stuart Coles. The functions may be divided into the following 
  groups; maxima/minima, order statistics, peaks over thresholds
  and point processes.  \pkg{ismev} is an R port of Stuart Coles'
  S-Plus extreme value statistical routines.

  Primary functions include: 'gev.fit', 'gev.diag', 'gpd.fit', 'gpd.diag',
	'pp.fit' and 'pp.diag'.

  Original R port was carried out by alec Stephenson, and the package is
  currently being maintained by Eric Gilleland.
}
\keyword{misc}
