\name{portpirie}
\alias{portpirie}
\title{Annual Maximum Sea Levels at Port Pirie, South Australia}
\usage{data(portpirie)}
\description{
  The \code{portpirie} data frame has 65 rows and 2 columns.
  The second column gives annual maximimum sea levels recorded
  at Port Pirie, South Australia, from 1923 to 1987.
  The first column gives the corresponding years.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Year}{A numeric vector of years.}
    \item{Sea Level}{A numeric vector of annual sea level
      maxima.}
  }
}
\source{
  Coles, S. G. (2001)
  \emph{An Introduction to Statistical Modelling of
    Extreme Values.} London: Springer.
}
\keyword{datasets}
