\name{rain}
\alias{rain}
\title{Daily Rainfall Accumulations in South-West England}
\usage{data(rain)}
\description{
  A numeric vector containing daily rainfall accumulations
  at a location in south-west England over the period 1914
  to 1962.
}
\format{A vector containing 17531 observations.}
\source{
  Coles, S. G. and Tawn, J. A. (1996)
  Modelling extremes of the areal rainfall process.
  \emph{Journal of the Royal Statistical Society, B}
  \bold{53}, 329--347.
}
\references{
  Coles, S. G. (2001)
  \emph{An Introduction to Statistical Modelling of
    Extreme Values.} London: Springer.
}

\keyword{datasets}
