\name{wavesurge}
\alias{wavesurge}
\title{Wave and Surge Heights in South-West England}
\usage{data(wavesurge)}
\description{
  The \code{wavesurge} data frame has 2894 rows and 2 columns.
  The columns contain wave and surge heights (in metres) at a
  single location off south-west England.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{wave}{A numeric vector of wave heights.}
    \item{surge}{A numeric vector of surge heights.}
  }
}
\source{
  Coles, S. G. (2001)
  \emph{An Introduction to Statistical Modelling of
    Extreme Values.} London: Springer.
}
\keyword{datasets}
