\name{wind}
\alias{wind}
\title{Annual Maximum Wind Speeds at Albany and Hartford}
\usage{data(wind)}
\description{
  The \code{wind} data frame has 40 rows and 3 columns.
  The second and third columns contain annual maximum wind
  speeds at Albany, New York and Hartford, Connecticut
  respectively, over the period 1944 to 1983.
  The first column gives the corresponding years.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Year}{A numeric vector of years.}
    \item{Hartford}{Annual maximum wind speeds at Hartford.}
    \item{Albany}{Annual maximum wind speeds at Albany.}
  }
}
\source{
  Coles, S. G. (2001)
  \emph{An Introduction to Statistical Modelling of
    Extreme Values.} London: Springer.
}

\keyword{datasets}
