\name{wooster}
\alias{wooster}
\title{Minimum Temperatures at Wooster, Ohio}
\usage{data(wooster)}
\description{
  A numeric vector containing daily minimum temperatures,
  in degrees Fahrenheit, at Wooster, Ohio, over the period
  1983 to 1988.
}
\format{A vector containing 1826 observations.}
\source{
  Coles, S. G., Tawn, J. A. and Smith, R. L. (1994)
  A seasonal Markov model for extremely low temperatures.
  \emph{Environmetrics} \bold{5}, 221--239.
}
\references{
  Coles, S. G. (2001)
  \emph{An Introduction to Statistical Modelling of
    Extreme Values.} London: Springer.

  Smith, R. L., Tawn, J. A. and Coles, S. G. (1997)
  Markov chain models for threshold exceedences.
  \emph{Biometrica} \bold{84}, 249--268.
}

\keyword{datasets}
