\name{ismev}
\alias{ismev}
\title{ismev -- an Introduction to Statistical Modeling of Extreme Values}
\description{
\pkg{ismev} includes functions to support the computations carried out in Coles (2001). The functions may be divided into the following groups; maxima/minima, order statistics, peaks over thresholds and point processes.  \pkg{ismev} is an R port of the S-Plus extreme value statistical routines believed to be originally written by Janet E. Heffernan.

Primary functions include:

\code{gev.fit}, \code{gev.diag}, \code{gpd.fit}, \code{gpd.diag}, \code{pp.fit} and \code{pp.diag}.

  Original R port was carried out by Alec G. Stephenson, and the package is
  currently being maintained by Eric Gilleland.

Datasets from Coles (2001) included are:

dowjones
euroex
fremantle
portpirie
venice
wind
engine
exchange
glass
rain
wavesurge
wooster

}
\references{
Coles, Stuart (2001) \emph{An Introduction to Statistical Modeling of Extreme Values}, London, UK: Springer, ISBN: 1852334592, 208 pp.
}
\keyword{misc}
